requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
python-dotenv==1.0.1
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
